# openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.17"
  },
  "host": "fzcjkosqguvhhllzmpyx.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/leads": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.leads.id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.name"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.source"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.status"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.revenue"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.marketing_interaction_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.updated_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/leads"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "leads"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.leads"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "leads"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.leads.id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.name"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.source"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.status"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.revenue"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.marketing_interaction_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.updated_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "leads"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.leads.id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.owner_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.session_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.normalized_email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.name"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.email"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.source"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.status"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.revenue"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.marketing_interaction_id"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.leads.updated_at"
          },
          {
            "$ref": "#/parameters/body.leads"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "leads"
        ]
      }
    },
...
  },
  "definitions": {
...
    "leads": {
      "required": [
        "id",
        "owner_id",
        "name",
        "source",
        "status",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "id": {
          "default": "gen_random_uuid()",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "uuid",
          "type": "string"
        },
        "owner_id": {
          "format": "uuid",
          "type": "string"
        },
        "session_id": {
          "format": "text",
          "type": "string"
        },
        "normalized_phone": {
          "format": "text",
          "type": "string"
        },
        "normalized_email": {
          "format": "text",
          "type": "string"
        },
        "name": {
          "format": "text",
          "type": "string"
        },
        "phone": {
          "format": "text",
          "type": "string"
        },
        "email": {
          "format": "text",
          "type": "string"
        },
        "source": {
          "format": "text",
          "type": "string"
        },
        "status": {
          "default": "new",
          "format": "text",
          "type": "string"
        },
        "revenue": {
          "default": 0,
          "format": "numeric",
          "type": "number"
        },
        "marketing_interaction_id": {
          "description": "Note:\nThis is a Foreign Key to `marketing_interactions.id`.<fk table='marketing_interactions' column='id'/>",
          "format": "uuid",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "updated_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.leads": {
      "name": "leads",
      "description": "leads",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/leads"
      }
    },
    "rowFilter.leads.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.owner_id": {
      "name": "owner_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.session_id": {
      "name": "session_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.normalized_phone": {
      "name": "normalized_phone",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.normalized_email": {
      "name": "normalized_email",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.name": {
      "name": "name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.phone": {
      "name": "phone",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.email": {
      "name": "email",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.source": {
      "name": "source",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.revenue": {
      "name": "revenue",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.marketing_interaction_id": {
      "name": "marketing_interaction_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.leads.updated_at": {
      "name": "updated_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}